+++ src/starplot/__init__.py
@@ -20,6 +20,6 @@ from .models import (
 from .styles import *  # noqa: F401 F403
 from .observer import Observer  # noqa: F401
 from .projections import *  # noqa: F401 F403
-
+from .config import settings  # noqa: F401
 
 from ibis import _  # noqa: F401 F403