+++ docs/changelog.md
@@ -5,7 +5,7 @@
 - Adds an equidistant projection
 - Introduces an Observer model to encapsulate a time and place
 - Makes zenith plots a separate plot type
-- Adds a function for plotting a star magnitude scale
+- Adds a function for plotting a star magnitude scale (experimental, feedback wanted!)
 - Adds more location options for legends and allows separate X/Y padding
 - Adds a `settings` class for managing global settings
 - Adds a global setting for how text is rendered in SVG exports